+++ Docs/Whats_New.md
@@ -19,6 +19,7 @@ MoltenVK 1.0.35
 Released TBD
 
 - Support the `VK_EXT_debug_report` extension.
+- Support the `VK_NV_glsl_shader` extension.
 - Change log indication of error in logs from `[***MoltenVK ERROR***]` to 
   `[mvk-error]`, for consistency with other log level indications. 
 - Tessellation fixes: